Use latest rate for recent transactions (#187)
- use latest rate from RateManager if it is not expired and transaction timestamp is within an hour
- change timestamp field for TransactionRecord from Int to Double

+++ BankWallet/BankWalletTests/TestExtensions.swift
@@ -41,7 +41,7 @@ extension TransactionFilterItem: Equatable {
 }
 
 extension TransactionRecord {
-    convenience init(transactionHash: String, coin: String, timestamp: Int) {
+    convenience init(transactionHash: String, coin: String, timestamp: Double) {
         self.init()
 
         self.transactionHash = transactionHash
